--- components/_util/getStatusClassNames.ts
@@ -1,16 +1,0 @@
-import classNames from 'classnames';
-import { ValidateStatus } from '../form/FormItem';
-
-export default function getStatusClassNames(
-  prefixCls: string,
-  status?: ValidateStatus,
-  hasFeedback?: boolean,
-) {
-  return classNames({
-    [`${prefixCls}-status-success`]: status === 'success',
-    [`${prefixCls}-status-warning`]: status === 'warning',
-    [`${prefixCls}-status-error`]: status === 'error',
-    [`${prefixCls}-status-validating`]: status === 'validating',
-    [`${prefixCls}-has-feedback`]: hasFeedback,
-  });
-}
