+++ components/input/TextArea.tsx
@@ -9,8 +9,7 @@ import { ConfigContext } from '../config-provider';
 import { fixControlledValue, resolveOnChange, triggerFocus, InputFocusOptions } from './Input';
 import SizeContext, { SizeType } from '../config-provider/SizeContext';
 import { ValidateStatus } from '../form/FormItem';
-import getStatusClassNames from '../_util/getStatusClassNames';
-import getFeedbackIcon from '../_util/getFeedbackIcon';
+import { getFeedbackIcon, getStatusClassNames } from '../_util/status-utils';
 import { FormItemStatusContext } from '../form/context';
 
 interface ShowCountProps {
